Decoupled firebase logic from AuthMiddleware's controller logic.
Moved admin.auth().verifyIdToken() logic (which includes checking user
roles) to services folder, under the auth.service.ts file, inside a
function called verifyRole().

AuthMiddleware imports that verifyRole() function and calls it to
check a user's role.
